Fixed bug where weapon image sometimes showed projectile instead of weapon.
Previous approach loaded animSheets in a map using name as a key,
for later retrieval during weaponAnim storage.

This approach is problematic because two different anims can share
the same name of an animSheet, but be referring to different sheets.

How? The animSheet used will be the closest one before the anim.
So now instead of loading animSheets in advance, during each
weaponAnim load, the matching animSheet is found first.