animSheet fix revision
The fix was finding all animSheets matching the weaponAnim sheet,
via an XPath query, then keeping each one until there was one
with an index after the weaponAnim, then halting.

We can actually just iterate backward starting from the
weaponAnim and grab the first animSheet with the name needed.

So the Jaxen dependency required for XPath can be removed,
and this should be faster overall: no query, just one search.