Ship txt file door entries are now sorted by X, Y, orientation
FTL doesn't care about door order in this file,
so door order in this file was random every save.

However, that just obfuscates the door info - if it can
be saved in a standard way, that makes it clear what has
changed (if anything), and makes the file all the more readable.