Added complete support for max level on player ships
Each systemBlueprint is now added to the database, enabling
the level slider to change with what's currently patched.

Systems on player ships now can toggle a max level slider,
which functions the same as when editing enemy ships.
When disabled, the starting level can be adjusted to the cap,
and the max level is set to the cap for future use.
The max level now loads and saves as expected for player ships.

Also fixes a bug with max level that was there for enemy ships:
switching from a system with a low max previously changed
the starting level of the new system down to that max:
for example, set weapons starting level to 5, switch to
medbay has a max 3, then weapons starting level is 3.