Added build for Mac arm64
Eclipse SWT apps have to be built for specific architectures.
Mac users with the newer Apple arm64 chip are therefore unable
to use the current Mac version which is for previous Intel x86-64.

Hence, this build is to enable those users to run Superluminal.
The library usage in this build is different from the others,
the reasoning is explained in its pom (modules/mac-arm64).