Fixed crash caused by missing systemBlueprint definitions
Each system object was getting added to the ship based on their
systemBlueprint definition. When vanilla is patched, this is
problematic for the Temporal system, creating a crash.

Now, any systems like this without systemBlueprint definitions
will have their levels guessed to be 1 to 3, as that is the
default Temporal as well as other non-essential systems.

Also added potential save crash fix and getSystems helper methods.