Fixed bug where lower system max affected start levels
When switching systems, checkbox for max level was being
updated before the level and max level values were
retrieved for the next system selected.

The current system's max was getting lowered to previous
system's max possible amount. This lower max in turn lowered
the start level too, as that cannot be higher than the max.

Essentially, the 3 components - two sliders and checkbox -
must be updated in the right order that they relate in.