Warn on ship load w/unknown hidden aug
Ships with hidden augs that couldn't be located would
error and just not load at all. The ship will load now
with a warning that it couldn't load its hidden aug(s).

This is either the first of such ship load fixes,
or a prelude to a placeholder mode where ships can load
with equipment not found in the current ftl.dat.